requirements.txt
beartype
python-dateutil
gspread
pre-commit
tabulate
rich
